Fix: Change txnum type to int64 in IEviction (1) ``` Change the type of txnum parameter in FlushAll method of IEviction interface from int to int64. ```

+++ buffer/IEviction.go
@@ -13,5 +13,5 @@ type EvictionPolicy interface {
 	// Evict removes a block from the buffer based on the eviction policy.
 	Evict() (*Buffer, error)
 
-	FlushAll(txnum int)
+	FlushAll(txnum int64)
 }
